+++ sha224.h
@@ -6,7 +6,7 @@
 /*   By: yez-zain <yez-zain@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/03 11:49:02 by yez-zain          #+#    #+#             */
-/*   Updated: 2022/05/05 11:47:31 by yez-zain         ###   ########.fr       */
+/*   Updated: 2022/05/05 14:27:07 by yez-zain         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@
 # include "ft_ssl.h"
 # include "sha32bits.h"
 
-char		*sha224(const char *str, uint32_t len);
+char		*sha224(const char *str, uint64_t len);
 char		*sha224_from_stream(int fd);
 void		sha224_init_ctx(t_sha32bits_context *ctx);
